- Use multiprocessing in cleaner

+++ scripts/clean.py
@@ -1,6 +1,5 @@
 #!/usr/bin/env python3
 import argparse
-import asyncio
 import json
 import os
 import shutil
@@ -9,44 +8,58 @@ from loguru import logger
 from tqdm import tqdm
 
 from src.cat.catter import Catter
-from src.configs.datasets import SPIDER_ALL
+from src.configs.config_loader import load_config
 from src.dataset.models import SpiderExample
-from src.rel.db_facade import DatabaseFactory
-from src.util.multi_thread_utils import exec_multi_thread
+from src.eval.dataset_config import DatasetConfig
+from src.rel.db_facade import DatabaseFactory, DatabaseFacade
+from src.util.log_util import configure_logging
+from src.util.multi_thread_utils import exec_multi_process_flat
 
 
-async def clean_dataset(file_path: str, db_path: str):
-    db_facade = DatabaseFactory.get_instance(SPIDER_ALL)
+class SqlExecWorker:
+    db_facade: DatabaseFacade
+
+    def __init__(self, config):
+        self.db_facade = DatabaseFactory.get_instance(config)
+
+    def exec(self, ex):
+        r = self.db_facade.exec_query_sync(ex.db_id, ex.query)
+        return r
+
+    def exec_for_examples(self, data):
+        res = []
+        for ex in tqdm(data, desc=f"[{os.getpid()}]: Executing queries", total=len(data)):
+            r = self.exec(ex)
+            res.append(r)
+        return res
+
+
+def clean_file(dataset_conf: DatasetConfig, data_file_path: str, overwrite: bool):
+    db_facade = DatabaseFactory.get_instance(dataset_conf)
     catter = Catter()
     errors = []
     total = 0
     valid_examples = []
-    data_file_path = file_path
     with open(data_file_path) as file:
         data = json.load(file)
+        # data = data[4000:4500]
         examples = []
         for i, entry in enumerate(data):
             example = SpiderExample.model_validate(entry)
             example.idx = i
             examples.append(example)
 
-        def exec_for_examples(data):
-            res = []
-            for ex in data:
-                r = db_facade.exec_query_sync(ex.db_id, ex.query)
-                res.append(r)
-            return res
-
-        results = exec_multi_thread(exec_for_examples, examples)
+        worker = SqlExecWorker(dataset_conf)
+        #
+        results = exec_multi_process_flat(worker.exec, examples)
+        # results = list(tqdm(map(worker.exec, examples),total=len(examples)))
 
         for i, entry in tqdm(enumerate(data), colour="green", total=len(data),
-                             desc=f"Validating dataset: {db_path}"):
+                             desc=f"Validating dataset: {dataset_conf.dataset_type}"):
             example = SpiderExample.model_validate(entry)
             cat = catter.get_category(example.query)
             exec_res = results[i]
-            if isinstance(exec_res, asyncio.TimeoutError):
-                errors.append((i, example.db_id, 'TIMEOUT', example.query))
-            elif exec_res is None:
+            if exec_res is None:
                 errors.append((i, example.db_id, 'EXEC_ERROR', example.query))
             elif cat is None:
                 errors.append((i, example.db_id, 'PARSE_ERROR', example.query))
@@ -68,8 +81,9 @@ async def clean_dataset(file_path: str, db_path: str):
         shutil.copy(data_file_path, f"{file_stem}.dirty.json")
         logger.info(f"Copied data file to {file_stem}.dirty.json")
 
-        with open(f"{data_file_path}", "w") as out_file:
-            out_file.write(json.dumps(valid_examples, indent=True))
+        if overwrite:
+            with open(f"{data_file_path}", "w") as out_file:
+                out_file.write(json.dumps(valid_examples, indent=True))
 
         with open(f"{file_stem}.errors.json", "w") as out_file:
             out_file.write(json.dumps(errors, indent=True))
@@ -79,13 +93,21 @@ async def clean_dataset(file_path: str, db_path: str):
         logger.info("Dataset is valid!")
 
 
-async def main(data_file, db_dir):
-    await clean_dataset(data_file, db_dir)
+def clean_dataset(config_path: str, overwrite: bool):
+    conf = load_config(config_path)
+    dataset_conf = conf.eval_conf.dataset_config
+    clean_file(dataset_conf, dataset_conf.get_train_path(), overwrite)
+    clean_file(dataset_conf, dataset_conf.get_test_path(), overwrite)
+
+
+def main(config_file, overwrite):
+    clean_dataset(config_file, overwrite)
 
 
 if __name__ == '__main__':
+    configure_logging()
     parser = argparse.ArgumentParser()
-    parser.add_argument("-f", required=True, help="Data file")
-    parser.add_argument("-d", required=True, help="Database dir")
+    parser.add_argument("-c", required=True, help="Config file")
+    parser.add_argument("-f", required=False, action='store_true', default=False, help="Overwrite")
     args = parser.parse_args()
-    asyncio.run(main(args.f, args.d))
+    main(args.c, args.f)
